requirements.txt
Django~=4.0.4
Pillow~=9.1.0
django-crispy-forms==1.11.1
python-dateutil~=2.8.2